moved boon data to AbstractMasterPlayer
