+++ LuckParser/Models/ParseModels/Players/AbstractPlayer.cs
@@ -11,10 +11,6 @@ namespace LuckParser.Models.ParseModels
     {
         protected AgentItem agent;
         private String character;
-        // Boons
-        private List<BoonDistribution> boon_distribution = new List<BoonDistribution>();
-        private List<Dictionary<int, long>> boon_presence = new List<Dictionary<int, long>>();
-        private Dictionary<int, BoonsGraphModel> boon_points = new Dictionary<int, BoonsGraphModel>();
         // DPS
         protected List<DamageLog> damage_logs = new List<DamageLog>();
         protected List<DamageLog> damage_logsFiltered = new List<DamageLog>();
@@ -97,30 +93,6 @@ namespace LuckParser.Models.ParseModels
             }
             return damageTaken_logs.Where(x => x.getTime() >= start && x.getTime() <= end).ToList();
         }
-        public BoonDistribution getBoonDistribution(ParsedLog log, List<PhaseData> phases, List<Boon> to_track, int phase_index)
-        {
-            if (boon_distribution.Count == 0)
-            {
-                setBoonDistribution(log, phases, to_track);
-            }
-            return boon_distribution[phase_index];
-        }
-        public Dictionary<int, BoonsGraphModel> getBoonGraphs(ParsedLog log, List<PhaseData> phases, List<Boon> to_track)
-        {
-            if (boon_distribution.Count == 0)
-            {
-                setBoonDistribution(log, phases, to_track);
-            }
-            return boon_points;
-        }
-        public Dictionary<int, long> getBoonPresence(ParsedLog log, List<PhaseData> phases, List<Boon> to_track, int phase_index)
-        {
-            if (boon_distribution.Count == 0)
-            {
-                setBoonDistribution(log,phases, to_track);
-            }
-            return boon_presence[phase_index];
-        }
         public List<CastLog> getCastLogs(ParsedLog log, long start, long end)
         {
             if (cast_logs.Count == 0)
@@ -189,206 +161,5 @@ namespace LuckParser.Models.ParseModels
         protected abstract void setFilteredLogs(ParsedLog log);
         protected abstract void setCastLogs(ParsedLog log);
         protected abstract void setDamagetakenLogs(ParsedLog log);
-
-        private static void Add<T>(Dictionary<T, long> dictionary, T key, long value)
-        {
-            if (dictionary.TryGetValue(key, out var existing))
-            {
-                dictionary[key] = existing + value;
-            }
-            else
-            {
-                dictionary.Add(key, value);
-            }
-        }
-
-        private void setBoonDistribution(ParsedLog log,List<PhaseData> phases, List<Boon> to_track)
-        {
-            BoonMap to_use = getBoonMap(log, to_track);
-            long dur = log.getBossData().getAwareDuration();
-            int fight_duration = (int)(dur) / 1000;
-            // Init boon presence points
-            BoonsGraphModel boon_presence_points = new BoonsGraphModel("Number of Boons");
-            for (int i = 0; i <= fight_duration; i++)
-            {
-                boon_presence_points.getBoonChart().Add(new Point(i, 0));
-            }
-            for (int i = 0; i < phases.Count; i++)
-            {
-                boon_distribution.Add(new BoonDistribution());
-                boon_presence.Add(new Dictionary<int, long>());
-            }
-
-            var toFill         = new Point[dur + 1];
-            var toFillPresence = new Point[dur + 1];
-
-            long death = getDeath(log, 0, dur);
-
-            foreach (Boon boon in to_track)
-            {
-                int boonid = boon.getID();
-                if (to_use.TryGetValue(boonid, out var logs) && logs.Count != 0)
-                {
-                    if (boon_distribution[0].ContainsKey(boonid))
-                    {
-                        continue;
-                    }
-                    var simulator = boon.CreateSimulator();
-                    simulator.simulate(logs, dur);
-                    if (death > 0 && getCastLogs(log, death + 1, fight_duration).Count > 0)
-                    {
-                        simulator.trim(death - log.getBossData().getFirstAware());
-                    }
-                    else
-                    {
-                        simulator.trim(dur);
-                    }
-                    var simulation = simulator.getSimulationResult();
-                    foreach (var simul in simulation)
-                    {
-                        for (int i = 0; i < phases.Count; i++)
-                        {
-                            var phase        = phases[i];
-                            var presenceDict = boon_presence[i];
-                            if (!boon_distribution[i].TryGetValue(boonid, out var distrib))
-                            {
-                                distrib = new Dictionary<ushort, OverAndValue>();
-                                boon_distribution[i].Add(boonid, distrib);
-                            }
-                            Add(presenceDict, boonid, simul.getItemDuration(phase.getStart(), phase.getEnd()));
-                            foreach (ushort src in simul.getSrc())
-                            {
-                                if (distrib.TryGetValue(src, out var toModify))
-                                {
-                                    toModify.value     += simul.getDuration (src, phase.getStart(), phase.getEnd());
-                                    toModify.overstack += simul.getOverstack(src, phase.getStart(), phase.getEnd());
-                                    distrib[src] = toModify;
-                                }
-                                else
-                                {
-                                    distrib.Add(src, new OverAndValue(
-                                        simul.getDuration (src, phase.getStart(), phase.getEnd()),
-                                        simul.getOverstack(src, phase.getStart(), phase.getEnd())));
-                                }
-                            }
-                        }
-                    }
-                    // Graphs
-                    // full precision
-                    for (int i = 0; i <= dur; i++)
-                    {
-                        toFill[i]         = new Point(i, 0);
-                        toFillPresence[i] = new Point(i, 0);
-                    }
-                    foreach (var simul in simulation)
-                    {
-                        int start = (int)simul.getStart();
-                        int end   = (int)simul.getEnd();
-
-                        bool present = simul.getItemDuration() > 0;
-                        for (int i = start; i <= end; i++)
-                        {
-                            toFill[i]         = new Point(i, simul.getStack(i));
-                            toFillPresence[i] = new Point(i, present ? 1 : 0);
-                        }
-                    }
-                    // reduce precision to seconds
-                    var reducedPrecision = new List<Point>(capacity: fight_duration + 1);
-                    var boonPresence = boon_presence_points.getBoonChart();
-                    var updateBoonPresence = Boon.getBoonList().Any(x => x.getID() == boonid);
-                    for (int i = 0; i <= fight_duration; i++)
-                    {
-                        reducedPrecision.Add(new Point(i, toFill[1000 * i].Y));
-                        if (updateBoonPresence)
-                        {
-                            boonPresence[i] = new Point(i, boonPresence[i].Y + toFillPresence[1000 * i].Y);
-                        }
-                    }
-                    boon_points[boonid] = new BoonsGraphModel(boon.getName(), reducedPrecision);
-                }
-            }
-            boon_points[-2] = boon_presence_points;
-        }
-
-        // private getters
-        private BoonMap getBoonMap(ParsedLog log, List<Boon> to_track)
-        {
-            BoonMap boon_map = new BoonMap();
-            boon_map.add(to_track);
-            // Fill in Boon Map
-            long time_start = log.getBossData().getFirstAware();
-
-            foreach (CombatItem c in log.getBoonData())
-            {
-                if (!boon_map.ContainsKey(c.getSkillID()))
-                {
-                    continue;
-                }
-                long time = c.getTime() - time_start;
-                ushort dst = c.isBuffremove() == ParseEnum.BuffRemove.None ? c.getDstInstid() : c.getSrcInstid();
-                if (agent.getInstid() == dst && time > 0 && time < log.getBossData().getAwareDuration())
-                {
-                    ushort src = c.getSrcMasterInstid() > 0 ? c.getSrcMasterInstid() : c.getSrcInstid();
-                    if (c.isBuffremove() == ParseEnum.BuffRemove.None)
-                    {
-                        boon_map[c.getSkillID()].Add(new BoonLog(time, src, c.getValue(), 0));
-                    }
-                    else if (Boon.removePermission(c.getSkillID(), c.isBuffremove(), c.getIFF()))
-                    {
-                        if (c.isBuffremove() == ParseEnum.BuffRemove.All)//All
-                        {
-                            List<BoonLog> loglist = boon_map[c.getSkillID()];
-                            for (int cnt = loglist.Count() - 1; cnt >= 0; cnt--)
-                            {
-                                BoonLog curBL = loglist[cnt];
-                                if (curBL.getTime() + curBL.getValue() > time)
-                                {
-                                    long subtract = (curBL.getTime() + curBL.getValue()) - time;
-                                    loglist[cnt].addValue(-subtract);
-                                    // add removed as overstack
-                                    loglist[cnt].addOverstack((ushort)subtract);
-                                }
-                            }
-
-                        }
-                        else if (c.isBuffremove() == ParseEnum.BuffRemove.Single)//Single
-                        {
-                            List<BoonLog> loglist = boon_map[c.getSkillID()];
-                            int cnt = loglist.Count() - 1;
-                            BoonLog curBL = loglist[cnt];
-                            if (curBL.getTime() + curBL.getValue() > time)
-                            {
-                                long subtract = (curBL.getTime() + curBL.getValue()) - time;
-                                loglist[cnt].addValue(-subtract);
-                                // add removed as overstack
-                                loglist[cnt].addOverstack((ushort)subtract);
-                            }
-                        }
-                        else if (c.isBuffremove() == ParseEnum.BuffRemove.Manual)//Manuel
-                        {
-                            List<BoonLog> loglist = boon_map[c.getSkillID()];
-                            for (int cnt = loglist.Count() - 1; cnt >= 0; cnt--)
-                            {
-                                BoonLog curBL = loglist[cnt];
-                                long ctime = curBL.getTime() + curBL.getValue();
-                                if (ctime > time)
-                                {
-                                    long subtract = (curBL.getTime() + curBL.getValue()) - time;
-                                    loglist[cnt].addValue(-subtract);
-                                    // add removed as overstack
-                                    loglist[cnt].addOverstack((ushort)subtract);
-                                    break;
-                                }
-                            }
-
-                        }
-                    }
-
-                }
-            }
-            return boon_map;
-        }
-        
-
     }
 }
